Fix wrong headers import
The headers to import in the customer app can't be changed otherwise it breaks the supported API and it requires to bump to the next major version.
