+++ Snowplow/Events/SPConsentDocument.h
@@ -6,7 +6,8 @@
 //  Copyright © 2020 Snowplow Analytics. All rights reserved.
 //
 
-#import "SPEvent.h"
+#import "SPEventBase.h"
+#import "SPSelfDescribingJson.h"
 
 NS_ASSUME_NONNULL_BEGIN
