+++ Snowplow/Events/SPEcommerceItem.h
@@ -6,7 +6,7 @@
 //  Copyright © 2020 Snowplow Analytics. All rights reserved.
 //
 
-#import "SPEvent.h"
+#import "SPEventBase.h"
 
 NS_ASSUME_NONNULL_BEGIN
