+++ Snowplow/Events/SPPushNotification.h
@@ -6,7 +6,10 @@
 //  Copyright © 2020 Snowplow Analytics. All rights reserved.
 //
 
-#import "SPEvent.h"
+#import "SPEventBase.h"
+#import "SPSelfDescribingJson.h"
+
+@class SPNotificationContent;
 
 NS_ASSUME_NONNULL_BEGIN
