+++ Snowplow/Events/SPScreenView.h
@@ -6,7 +6,10 @@
 //  Copyright © 2020 Snowplow Analytics. All rights reserved.
 //
 
-#import "SPEvent.h"
+#import "SPEventBase.h"
+#import "SPSelfDescribingJson.h"
+
+@class SPScreenState;
 
 NS_ASSUME_NONNULL_BEGIN
