+++ Snowplow/UIViewController+SPScreenView_SWIZZLE.m
@@ -7,7 +7,8 @@
 //
 
 #import "SPTracker.h"
-#import "SPEvent.h"
+#import "SPEventBase.h"
+#import "SPSelfDescribingJson.h"
 #import "SPUtilities.h"
 #import "UIKit/UIKit.h"
 #import "UIViewController+SPScreenView_SWIZZLE.h"
