+++ Snowplow/ios.modulemap
@@ -9,7 +9,7 @@ framework module SnowplowTracker {
     header "SPSelfDescribingJson.h"
     header "SPDevicePlatform.h"
     
-    header "SPEvent.h"
+    header "SPEventBase.h"
     header "SPPageView.h"
     header "SPStructured.h"
     header "SPUnstructured.h"
